# server/generated/openapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "prompton-api",
    "description": "API for prompton - managing full lifecycle of AI chat prompts.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/token": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Get Access Token Extended",
        "operationId": "get_access_token_extended",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_get_access_token_extended_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect username or password"
          },
          "400": {
            "description": "Malformed request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/token_basic": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Get Access Token",
        "description": "Same functionality as /token but taking `username` and `password` args as `application/json` type instead of `application/x-www-form-urlencoded`\n\nDon't use it becuase it's a **temporary** workaround for client lib generator and will be removed in the future.",
        "operationId": "get_access_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCredentialsPost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect username or password"
          },
          "400": {
            "description": "Malformed request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prompts": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Prompts List",
        "operationId": "get_prompts_list",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Prompts List Prompts Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PromptRead"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid itemid supplied"
          },
          "404": {
            "description": "Item not found"
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "prompts"
        ],
        "summary": "Add Prompt",
        "operationId": "add_prompt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Item created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultPostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/prompts/{id}": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Prompt By Id",
        "operationId": "get_prompt_by_id",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptRead"
                }
              }
            }
          },
          "422": {
            "description": "Invalid itemid supplied"
          },
          "404": {
            "description": "Item not found"
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "prompts"
        ],
        "summary": "Delete Prompt",
        "description": "Not implemented",
        "operationId": "delete_prompt",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "405": {
            "description": "Method not implemented",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Invalid itemid supplied"
          },
          "404": {
            "description": "Item not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "patch": {
        "tags": [
          "prompts"
        ],
        "summary": "Update Prompt",
        "operationId": "update_prompt",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Item updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptRead"
                }
              }
            }
          },
          "422": {
            "description": "Invalid itemid supplied"
          },
          "404": {
            "description": "Item not found"
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/feedbacks": {
      "get": {
        "tags": [
          "feedbacks"
        ],
        "summary": "Get Feedbacks List",
        "operationId": "get_feedbacks_list",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Prompt Version Id",
              "type": "string"
            },
            "name": "prompt_version_id",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Prompton User Id",
              "type": "string"
            },
            "name": "prompton_user_id",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Inference Id",
              "type": "string"
            },
            "name": "inference_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Feedbacks List Feedbacks Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedbackRead"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid itemid supplied"
          },
          "404": {
            "description": "Item not found"
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "feedbacks"
        ],
        "summary": "Add Feedback",
        "description": "Feedback about an inference from an end user or from a PromptOn user.\n\nAt least one of `score`, `flag` or `note` required\n\n`end_user_id` is null if from PromptOn user.\n\n`score` can be null if flagging or note only. Scoring method are for API consumer consideration. Eg. end users can thumbs down/up mapped to -1 +1 `score`, internal users rate between from 0 to 100.\n\n`flag` is any string when inference was flagged. Can be null if it is scoring or note only\n\n`feedback_for_part` is optional and can be used when the inference has multiple sections which require separate feedback.\n\n`metadata` is any dict for additional information.",
        "operationId": "add_feedback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Item created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultPostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/feedbacks/{id}": {
      "get": {
        "tags": [
          "feedbacks"
        ],
        "summary": "Get Feedback By Id",
        "operationId": "get_feedback_by_id",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackRead"
                }
              }
            }
          },
          "422": {
            "description": "Invalid itemid supplied"
          },
          "404": {
            "description": "Item not found"
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_get_access_token_extended_token_post": {
        "title": "Body_get_access_token_extended_token_post",
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "properties": {
          "grant_type": {
            "title": "Grant Type",
            "pattern": "password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string",
            "default": ""
          },
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "client_secret": {
            "title": "Client Secret",
            "type": "string"
          }
        }
      },
...
      "DefaultPostResponse": {
        "title": "DefaultPostResponse",
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string",
            "description": "The id of the created item"
          }
        },
        "description": "Standard response for POST requests. Contains the id of the created item."
      },
      "FeedbackCreate": {
        "title": "FeedbackCreate",
        "required": [
          "inference_id"
        ],
        "type": "object",
        "properties": {
          "inference_id": {
            "title": "Inference Id",
            "type": "string",
            "description": "The inference being rated"
          },
          "end_user_id": {
            "title": "End User Id",
            "type": "string",
            "description": "API consumers' end user id If feedback from end user otherwise null"
          },
          "feedback_for_part": {
            "title": "Feedback For Part",
            "type": "string",
            "description": "Specifies which part of the output the feedback is about. Can be used when the inference has multiple sections which require separate feedback"
          },
          "score": {
            "title": "Score",
            "type": "integer",
            "description": "Any integer score. Rules are up to the API consumer. Can be null if it was flagging or note only"
          },
          "flag": {
            "title": "Flag",
            "type": "string",
            "description": "Any string when inference was flagged. Can be null if it is scoring or note only"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          }
        },
        "additionalProperties": false
      },
      "FeedbackRead": {
        "title": "FeedbackRead",
        "required": [
          "_id",
          "created_at",
          "created_by_user_id",
          "created_by_org_id",
          "inference_id"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          },
          "created_by_user_id": {
            "title": "Created By User Id",
            "type": "string"
          },
          "created_by_org_id": {
            "title": "Created By Org Id",
            "type": "string"
          },
          "inference_id": {
            "title": "Inference Id",
            "type": "string",
            "description": "The inference being rated"
          },
          "end_user_id": {
            "title": "End User Id",
            "type": "string",
            "description": "API consumers' end user id If feedback from end user otherwise null"
          },
          "feedback_for_part": {
            "title": "Feedback For Part",
            "type": "string",
            "description": "Specifies which part of the output the feedback is about. Can be used when the inference has multiple sections which require separate feedback"
          },
          "score": {
            "title": "Score",
            "type": "integer",
            "description": "Any integer score. Rules are up to the API consumer. Can be null if it was flagging or note only"
          },
          "flag": {
            "title": "Flag",
            "type": "string",
            "description": "Any string when inference was flagged. Can be null if it is scoring or note only"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "prompt_version_id": {
            "title": "Prompt Version Id",
            "type": "string"
          }
        },
        "description": "Base model for reading from MongoDB. Same as MongoBaseCreate but assumes all DB base fields are populated so generated clients doesn't requrie None checks"
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "LoginCredentialsPost": {
        "title": "LoginCredentialsPost",
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "title": "Username",
            "type": "string",
            "format": "email"
          },
          "password": {
            "title": "Password",
            "type": "string",
            "format": "password",
            "writeOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "PromptCreate": {
        "title": "PromptCreate",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PromptStatus"
              }
            ],
            "description": "Prompt status for client consideration only, currently not used in server logic.",
            "default": "Active"
          },
          "name": {
            "title": "Name",
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PromptRead": {
        "title": "PromptRead",
        "required": [
          "_id",
          "created_at",
          "created_by_user_id",
          "created_by_org_id",
          "status",
          "name"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          },
          "created_by_user_id": {
            "title": "Created By User Id",
            "type": "string"
          },
          "created_by_org_id": {
            "title": "Created By Org Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PromptStatus"
          },
          "name": {
            "title": "Name",
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          }
        },
        "description": "Base model for reading from MongoDB. Same as MongoBaseCreate but assumes all DB base fields are populated so generated clients doesn't requrie None checks"
      },
      "PromptStatus": {
        "title": "PromptStatus",
        "enum": [
          "Active",
          "Archived"
        ],
        "type": "string",
        "description": "An enumeration."
      },
      "PromptUpdate": {
        "title": "PromptUpdate",
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PromptStatus"
              }
            ],
            "description": "Prompt status for client consideration only, currently not used in server logic.",
            "default": "Active"
          },
          "name": {
            "title": "Name",
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Token": {
        "title": "Token",
        "required": [
          "access_token",
          "token_type"
        ],
        "type": "object",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "title": "Token Type",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}